requirements.txt
playwright==1.39.0
google-generativeai
chromadb
pydub
SpeechRecognition
openai-whisper
